format code
